@@ -1,7 +1,7 @@
 # https://black.readthedocs.io/
 [tool.black]
 line-length = 88
-target_version = ['py36']
+target_version = ['py38']
 include = '\.pyi?$'
 exclude = '''
 /(
