@@ -356,7 +356,7 @@ class AdvertiserViewSet(viewsets.ReadOnlyModelViewSet):
         return self.request.user.advertisers.all()
 
     @action(detail=True, methods=["get"])
-    def report(self, request, slug=None):
+    def report(self, request, slug=None):  # pylint: disable=unused-argument
         """Return a report of ad performance for this advertiser."""
         # This will raise a 404 if the user doesn't have access to the advertiser
         advertiser = self.get_object()
@@ -465,7 +465,7 @@ class PublisherViewSet(viewsets.ReadOnlyModelViewSet):
         return self.request.user.publishers.all()
 
     @action(detail=True, methods=["get"])
-    def report(self, request, slug=None):
+    def report(self, request, slug=None):  # pylint: disable=unused-argument
         """Return a report of ad performance for this publisher."""
         # This will raise a 404 if the user doesn't have access to the publisher
         publisher = self.get_object()
